Fixed AndroidHorizontalScrollView issue
react-native 0.49.0 introduces a weird breaking change that don't allow
us to use requireNativeComponent twice of the same view name, and since
we are using the same implementation of ScrollView.js, we are requiring
AndroidHorizontalScrollView twice, which now conflicts with the current
ScrollView component and throws a error.

This commit implements ReactHorizontalNestedScrollViewManager that
replaces the native view name (REACT_CLASS) and extends the default behavior
of the ReactHorizontalScrollViewManager.